removed workers and cleaned up more
