+++ base.yml
@@ -15,13 +15,10 @@ services:
       - "BAT_BALANCE_SERVER=http://balance-web:3003"
       - "BAT_EYESHADE_SERVER=http://eyeshade-web:3002"
       - "BAT_GRANT_POSTGRES_URL=postgres://grants:password@grant-postgres/grants?sslmode=disable"
-      - "BAT_GRANT_SERVER=http://grant-web:3333"
       - "BAT_WALLET_MIGRATION_SERVER=http://wallet-migration-web:3333"
       - "WALLET_MIGRATION_SERVER=http://wallet-migration-web:3333"
       - "WALLET_MIGRATION_TOKEN=${WALLET_MIGRATION_TOKEN-98f27f2e-239b-40e9-8f62-bf1a55ea530c}"
       - "BAT_WALLET_MIGRATION_POSTGRES_URL=${BAT_WALLET_MIGRATION_POSTGRES_URL-postgres://wallet-migration:password@wallet-migration-postgres/wallet-migration}"
-      - "BAT_LEDGER_SERVER=http://ledger-web:3001"
-      - "BAT_MONGODB_URI=mongodb://mongo"
       - "BAT_POSTGRES_URL=postgres://eyeshade:password@eyeshade-postgres/eyeshade"
       - "BAT_RATIOS_URL=https://ratios-staging.mercury.basicattentiontoken.org"
       - "BAT_REDEEMER_REDIS_URL=redis://redeemer-redis:6379"
@@ -40,9 +37,6 @@ services:
       - "KAFKA_REQUIRED_ACKS=1"
       - "FEATURE_WALLET=true"
       - "FREEZE_SURVEYORS_AGE_DAYS=${FREEZE_SURVEYORS_AGE_DAYS-7}"
-      - "GRANT_SERVER=http://grant-web:3333"
-      - "LEDGER_URL=${LEDGER_URL-http://ledger-web:3001}"
-      - "MONGODB_URI=${MONGODB_URI-mongodb://mongo/ledger}"
       - "NODE_ENV=${NODE_ENV-development}"
       - "PER_MEMBER_CLAIM_LIMIT=4"
       - "REDEEMER_REDIS_HOST=${REDEEMER_REDIS_HOST-redeemer-redis}"
@@ -74,26 +68,13 @@ services:
       - COINBASE_WIDGET_CODE
       - DISABLE_WALLET_TO_GRANTS
       - DYNO
-      - EYESHADE_GITHUB_CLIENT_ID
-      - EYESHADE_GITHUB_CLIENT_SECRET
       - EYESHADE_SLACK_CHANNEL
       - EYESHADE_SLACK_ICON_URL
       - FIXIE_URL
       - FORWARD_TO_GRANTS
       - FORWARD_WALLET_TO_GRANTS
-      - GITHUB_CLIENT_ID
-      - GITHUB_CLIENT_SECRET
-      - GITHUB_LOGIN_BYE
-      - GITHUB_LOGIN_WORLD
-      - GITHUB_ORG
       - HOST
       - IP_GRAYLIST
-      - LEDGER_GITHUB_CLIENT_ID
-      - LEDGER_GITHUB_CLIENT_SECRET
-      - LEDGER_SLACK_CHANNEL
-      - LEDGER_SLACK_ICON_URL
-      - MIGRATE_DATABASE_URL
-      - MIGRATE_MONGODB_URI
       - PORT
       - PUBLISHERS_TOKEN
       - PUBLISHERS_URL
