+++ bat-utils/boot-extras.js
@@ -2,10 +2,8 @@
 const hapi = require('./lib/extras-hapi')
 const joi = require('./lib/extras-joi')
 const utils = require('./lib/extras-utils')
-const worker = require('./lib/extras-worker')
 module.exports = {
   joi,
   hapi,
-  utils,
-  worker
+  utils
 }
