+++ bat-utils/boot-runtime.js
@@ -11,9 +11,7 @@ const kafka = require('./lib/runtime-kafka')
 const newrelic = require('./lib/runtime-newrelic')
 const postgres = require('./lib/runtime-postgres')
 const prometheus = require('./lib/runtime-prometheus')
-const queue = require('./lib/runtime-queue')
 const sentry = require('./lib/runtime-sentry')
-const slack = require('./lib/runtime-slack')
 const wallet = require('./lib/runtime-wallet')
 const wreck = require('./lib/runtime-wreck')
 
@@ -25,9 +23,7 @@ const hash = {
   newrelic,
   postgres,
   prometheus,
-  queue,
   sentry,
-  slack,
   wallet,
   wreck
 }
