+++ eyeshade/consumer.js
@@ -22,7 +22,6 @@ Runtime.newrelic.setupNewrelic(config, __filename)
 
 config.cache = false
 config.database = false
-config.queue = false
 config.prometheus = false
 config.postgres.schemaVersion = require('./migrations/current')
