+++ eyeshade/server.js
@@ -11,7 +11,6 @@ const accountsController = require('./controllers/accounts')
 const publishersController = require('./controllers/publishers')
 const referralsController = require('./controllers/referrals')
 const statsController = require('./controllers/stats')
-const snapshotsController = require('./controllers/snapshots')
 
 Runtime.newrelic.setupNewrelic(config, __filename)
 
@@ -19,8 +18,7 @@ const parentModules = [
   accountsController,
   publishersController,
   referralsController,
-  statsController,
-  snapshotsController
+  statsController
 ]
 
 const options = {
@@ -33,7 +31,6 @@ const options = {
   remoteP: true
 }
 
-config.cache = false
 config.postgres.schemaVersion = require('./migrations/current')
 
 module.exports = server(options, new Runtime(config))
