+++ eyeshade/workers/reports.js
@@ -2,10 +2,9 @@ const moment = require('moment')
 const {
   timeout
 } = require('bat-utils/lib/extras-utils')
+const { surveyorFrozenReport } = require('./surveyors')
 
 const freezeInterval = process.env.FREEZE_SURVEYORS_AGE_DAYS
-
-const feePercent = 0.05
 
 const daily = async (debug, runtime) => {
   debug('daily', 'running')
@@ -62,7 +61,7 @@ async function freezeOldSurveyors (debug, runtime, olderThanDays) {
   const toFreeze = nonVirtualSurveyors.concat(virtualSurveyors)
   for (let i = 0; i < toFreeze.length; i += 1) {
     const surveyorId = toFreeze[i].id
-    await runtime.queue.send(debug, 'surveyor-frozen-report', { surveyorId, mix: true })
+    await surveyorFrozenReport(debug, runtime, { surveyorId, mix: true })
     await waitForTransacted(runtime, surveyorId)
   }
 }
@@ -71,7 +70,9 @@ async function waitForTransacted (runtime, surveyorId) {
   let row
   const start = new Date()
   do {
-    await timeout(5 * 1000)
+    if (+(new Date()) >= 5) {
+      await timeout(5 * 1000)
+    }
     const statement = `
     select *
     from votes
@@ -91,24 +92,6 @@ async function waitForTransacted (runtime, surveyorId) {
     }
   } while (row) // when no row is returned, all votes have been transacted
 }
-
-const mixer = async (runtime, client, surveyorId) => {
-  const query = `
-  update votes
-  set
-    amount = (1 - $1::decimal) * votes.tally * surveyor_groups.price,
-    fees =  $1::decimal * votes.tally * surveyor_groups.price
-  from surveyor_groups
-  where
-      votes.surveyor_id = surveyor_groups.id
-  and votes.surveyor_id = $2
-  and not votes.excluded
-  and surveyor_groups.frozen
-  `
-  return runtime.postgres.query(query, [feePercent, surveyorId], client)
-}
-
-exports.mixer = mixer
 
 exports.initialize = async (debug, runtime) => {
   if (typeof freezeInterval === 'undefined' || isNaN(parseFloat(freezeInterval))) {
