+++ package.json
@@ -94,9 +94,8 @@
     "proxy-agent": "^2.3.1",
     "rate-limiter-flexible": "^1.3.2",
     "redis": "^2.8.0",
-    "rsmq": "^0.8.2",
-    "rsmq-worker": "^0.5.2",
     "sdebug": "^0.9.5",
+    "sinek": "^7.30.1",
     "tldjs": "2.2.0",
     "tweetnacl": "^1.0.3",
     "underscore": "1.8.3",
