+++ test/isup.integration.test.js
@@ -6,18 +6,10 @@ const {
 
 test('check endpoint is up with no authorization', async (t) => {
   const {
-    BAT_BALANCE_SERVER,
-    BAT_EYESHADE_SERVER,
-    BAT_LEDGER_SERVER,
-    BAT_REDEEMER_SERVER,
-    BAT_GRANT_SERVER
+    BAT_EYESHADE_SERVER
   } = process.env
 
-  await checkIsUp(BAT_BALANCE_SERVER, 'ack.')
   await checkIsUp(BAT_EYESHADE_SERVER, 'ack.')
-  await checkIsUp(BAT_LEDGER_SERVER, 'ack.')
-  await checkIsUp(BAT_GRANT_SERVER, '.')
-  await checkIsUp(BAT_REDEEMER_SERVER, '.')
 
   async function checkIsUp (origin, expectation) {
     const {
